+++ backend/schedulerd/check_timer.go
@@ -82,17 +82,13 @@ type CronTimer struct {
 
 // NewCronTimer establishes new check timer given a name & an initial interval
 func NewCronTimer(name string, cronStr string) *CronTimer {
-	schedule, err := cron.ParseStandard(cronStr)
+	diff, err := NextCronTime(time.Now(), cronStr)
 	// we shouldn't hit this error because we've already validated the cron string
 	// but log and exit cleanly to revert to the interval timer
 	if err != nil {
 		logger.WithError(err).Error("invalid cron, reverting to interval")
 		return nil
 	}
-
-	nowTime := time.Now()
-	nextTime := schedule.Next(nowTime)
-	diff := nextTime.Sub(nowTime)
 	timer := &CronTimer{next: diff}
 	return timer
 }
@@ -104,16 +100,13 @@ func (timerPtr *CronTimer) C() <-chan time.Time {
 
 // SetDuration updates the interval in which timers are set
 func (timerPtr *CronTimer) SetDuration(cronStr string, interval uint) {
-	schedule, err := cron.ParseStandard(cronStr)
+	diff, err := NextCronTime(time.Now(), cronStr)
 	// we shouldn't hit this error because we've already validated the cron string
 	// but log and exit cleanly to revert to the interval timer
 	if err != nil {
 		logger.WithError(err).Error("invalid cron, reverting to interval")
 		return
 	}
-	nowTime := time.Now()
-	nextTime := schedule.Next(nowTime)
-	diff := nextTime.Sub(nowTime)
 	timerPtr.next = diff
 }
 
@@ -130,4 +123,17 @@ func (timerPtr *CronTimer) Next() {
 // Stop ends the timer
 func (timerPtr *CronTimer) Stop() bool {
 	return timerPtr.timer.Stop()
+}
+
+// NextCronTime calculates how much time is between the current time and the
+// time indidcated by the cron string
+func NextCronTime(now time.Time, cronStr string) (time.Duration, error) {
+	schedule, err := cron.ParseStandard(cronStr)
+	if err != nil {
+		return 0, err
+	}
+	nextTime := schedule.Next(now)
+	diff := nextTime.Sub(now)
+
+	return diff, nil
 }