+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Changed
 - Govaluate logic is now wrapped in the `util/eval` package.
+- Added functions for publishing proxy check requests.
+- Added proxy request validation.
 
 ## [2.0.0-alpha.14] - 2018-01-23
 ### Added